v0.1.2
Added the ability to create Cards using a class.

Card creation and deletion have animations.

Created a CardItem.js class template, content is yet to be coded.

index.js is used as an experimental markup to model the classes from.

Minor Bugs:
Visual bug in card deletion animation when scroll is overflowed.